Refactor to logcal ORM files
