+++ vectordb_orm/fields.py
@@ -1,0 +1,4 @@
+
+class Embedding:
+    def __init__(self, values: list[float]):
+        self.values = values
